# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Transcription & Translation App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface UserInfo {
  id: number;
  username: string;
  isAdmin: boolean;
  isActive: boolean;
  trialStartedAt?: string;
  trialEndsAt?: string;
  trialDaysRemaining: number;
  trialExpired: boolean;
  dailyLimitMinutes: number;
  minutesUsedToday: number;
  minutesRemainingToday: number;
  totalMinutesUsed: number;
  totalSessions: number;
}

export interface LoginResponse {
  user: UserInfo;
}

export interface TranscriptionTokenResponse {
  apiKey: string;
  expiresIn: number;
}

export interface SessionResponse {
  sessionId: number;
  message: string;
}

export interface StopSessionRequest {
  sessionId: number;
  durationSeconds: number;
}

export interface UsageInfo {
  minutesUsedToday: number;
  minutesRemainingToday: number;
  totalMinutesUsed: number;
  totalSessions: number;
  dailyLimitMinutes: number;
  trialDaysRemaining: number;
  trialExpired: boolean;
}

export interface FeedbackRequest {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string;
}

export interface AdminUser {
  id: number;
  username: string;
  isAdmin: boolean;
  isActive: boolean;
  trialStartedAt?: string;
  trialEndsAt?: string;
  dailyLimitMinutes: number;
  minutesUsedToday: number;
  totalMinutesUsed: number;
  totalSessions: number;
  createdAt: string;
}

export interface AdminUserListResponse {
  users: AdminUser[];
}

export interface AdminCreateUserRequest {
  username: string;
  password: string;
  isAdmin?: boolean;
  dailyLimitMinutes?: number;
}

export interface AdminUpdateUserRequest {
  isActive?: boolean;
  isAdmin?: boolean;
  dailyLimitMinutes?: number;
  password?: string;
}

export interface AdminFeedbackItem {
  id: number;
  userId: number;
  username: string;
  rating: number;
  comment?: string;
  createdAt: string;
}

export interface AdminFeedbackListResponse {
  feedback: AdminFeedbackItem[];
}