# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Transcription & Translation App API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with username and password
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    isAdmin: zod.boolean(),
    isActive: zod.boolean(),
    trialStartedAt: zod.coerce.date().optional(),
    trialEndsAt: zod.coerce.date().optional(),
    trialDaysRemaining: zod.number(),
    trialExpired: zod.boolean(),
    dailyLimitMinutes: zod.number(),
    minutesUsedToday: zod.number(),
    minutesRemainingToday: zod.number(),
    totalMinutesUsed: zod.number(),
    totalSessions: zod.number(),
  }),
});

/**
 * @summary Logout current user
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user info
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  isAdmin: zod.boolean(),
  isActive: zod.boolean(),
  trialStartedAt: zod.coerce.date().optional(),
  trialEndsAt: zod.coerce.date().optional(),
  trialDaysRemaining: zod.number(),
  trialExpired: zod.boolean(),
  dailyLimitMinutes: zod.number(),
  minutesUsedToday: zod.number(),
  minutesRemainingToday: zod.number(),
  totalMinutesUsed: zod.number(),
  totalSessions: zod.number(),
});

/**
 * @summary Get a short-lived Soniox API token for transcription
 */
export const GetTranscriptionTokenResponse = zod.object({
  apiKey: zod.string(),
  expiresIn: zod.number(),
});

/**
 * @summary Start a transcription session
 */
export const StartSessionResponse = zod.object({
  sessionId: zod.number(),
  message: zod.string(),
});

/**
 * @summary Stop the current transcription session
 */
export const StopSessionBody = zod.object({
  sessionId: zod.number(),
  durationSeconds: zod.number(),
});

export const StopSessionResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get usage stats for the current user
 */
export const GetMyUsageResponse = zod.object({
  minutesUsedToday: zod.number(),
  minutesRemainingToday: zod.number(),
  totalMinutesUsed: zod.number(),
  totalSessions: zod.number(),
  dailyLimitMinutes: zod.number(),
  trialDaysRemaining: zod.number(),
  trialExpired: zod.boolean(),
});

/**
 * @summary Submit feedback (trial end or general)
 */
export const submitFeedbackBodyRatingMax = 5;

export const SubmitFeedbackBody = zod.object({
  rating: zod.number().min(1).max(submitFeedbackBodyRatingMax),
  comment: zod.string().optional(),
});

export const SubmitFeedbackResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List all users (admin only)
 */
export const AdminListUsersResponse = zod.object({
  users: zod.array(
    zod.object({
      id: zod.number(),
      username: zod.string(),
      isAdmin: zod.boolean(),
      isActive: zod.boolean(),
      trialStartedAt: zod.coerce.date().optional(),
      trialEndsAt: zod.coerce.date().optional(),
      dailyLimitMinutes: zod.number(),
      minutesUsedToday: zod.number(),
      totalMinutesUsed: zod.number(),
      totalSessions: zod.number(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Create a new user (admin only)
 */
export const AdminCreateUserBody = zod.object({
  username: zod.string(),
  password: zod.string(),
  isAdmin: zod.boolean().optional(),
  dailyLimitMinutes: zod.number().optional(),
});

/**
 * @summary Update a user (admin only)
 */
export const AdminUpdateUserParams = zod.object({
  userId: zod.coerce.number(),
});

export const AdminUpdateUserBody = zod.object({
  isActive: zod.boolean().optional(),
  isAdmin: zod.boolean().optional(),
  dailyLimitMinutes: zod.number().optional(),
  password: zod.string().optional(),
});

export const AdminUpdateUserResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  isAdmin: zod.boolean(),
  isActive: zod.boolean(),
  trialStartedAt: zod.coerce.date().optional(),
  trialEndsAt: zod.coerce.date().optional(),
  dailyLimitMinutes: zod.number(),
  minutesUsedToday: zod.number(),
  totalMinutesUsed: zod.number(),
  totalSessions: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a user (admin only)
 */
export const AdminDeleteUserParams = zod.object({
  userId: zod.coerce.number(),
});

export const AdminDeleteUserResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Reset daily usage for a user (admin only)
 */
export const AdminResetUsageParams = zod.object({
  userId: zod.coerce.number(),
});

export const AdminResetUsageResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List all feedback (admin only)
 */
export const AdminListFeedbackResponse = zod.object({
  feedback: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.number(),
      username: zod.string(),
      rating: zod.number(),
      comment: zod.string().optional(),
      createdAt: zod.coerce.date(),
    }),
  ),
});