# lib/api-zod/src/generated/types/adminFeedbackItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Transcription & Translation App API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminFeedbackItem {
  id: number;
  userId: number;
  username: string;
  rating: number;
  comment?: string;
  createdAt: Date;
}