# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Transcription & Translation App API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  username: string;
  isAdmin: boolean;
  isActive: boolean;
  trialStartedAt?: Date;
  trialEndsAt?: Date;
  dailyLimitMinutes: number;
  minutesUsedToday: number;
  totalMinutesUsed: number;
  totalSessions: number;
  createdAt: Date;
}