# lib/api-zod/src/generated/types/stopSessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Transcription & Translation App API
 * OpenAPI spec version: 0.1.0
 */

export interface StopSessionRequest {
  sessionId: number;
  durationSeconds: number;
}