# lib/api-zod/src/generated/types/usageInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Transcription & Translation App API
 * OpenAPI spec version: 0.1.0
 */

export interface UsageInfo {
  minutesUsedToday: number;
  minutesRemainingToday: number;
  totalMinutesUsed: number;
  totalSessions: number;
  dailyLimitMinutes: number;
  trialDaysRemaining: number;
  trialExpired: boolean;
}